+++ src/lib/Makefile.inc
@@ -28,7 +28,10 @@ CSOURCES = ares__addrinfo2hostent.c	\
   ares_dns_parse.c			\
   ares_dns_record.c			\
   ares_dns_write.c			\
+  ares_event_kqueue.c			\
+  ares_event_poll.c			\
   ares_event_thread.c			\
+  ares_event_wake_pipe.c		\
   ares_expand_name.c			\
   ares_expand_string.c			\
   ares_fds.c				\
@@ -90,6 +93,7 @@ HHEADERS = ares__buf.h			\
   ares_android.h			\
   ares_data.h				\
   ares_dns_private.h			\
+  ares_event.h				\
   ares_getenv.h				\
   ares_inet_net_pton.h			\
   ares_ipv6.h				\
