+++ src/lib/ares_event_thread.c
@@ -26,106 +26,7 @@
 #include "ares_setup.h"
 #include "ares.h"
 #include "ares_private.h"
-#ifdef HAVE_POLL_H
-#  include <poll.h>
-#endif
-#ifdef HAVE_UNISTDD_H
-#  include <unistd.h>
-#endif
-#ifdef HAVE_FCNTL_H
-#  include <fcntl.h>
-#endif
-#ifdef HAVE_SYS_TYPES_H
-#  include <sys/types.h>
-#endif
-#ifdef HAVE_SYS_EVENT_H
-#  include <sys/event.h>
-#endif
-#ifdef HAVE_SYS_TIME_H
-#  include <sys/time.h>
-#endif
-
-struct ares_event;
-typedef struct ares_event ares_event_t;
-
-typedef enum {
-  ARES_EVENT_FLAG_NONE       = 0,
-  ARES_EVENT_FLAG_READ       = 1 << 0,
-  ARES_EVENT_FLAG_WRITE      = 1 << 1,
-  ARES_EVENT_FLAG_OTHER      = 1 << 2
-} ares_event_flags_t;
-
-typedef void (*ares_event_cb_t)(ares_event_thread_t *e, ares_socket_t fd,
-                                void *data, ares_event_flags_t flags);
-
-typedef void (*ares_event_free_data_t)(void *data);
-
-typedef void (*ares_event_signal_cb_t)(const ares_event_t *event);
-
-
-struct ares_event {
-  /*! Registered event thread this event is bound to */
-  ares_event_thread_t    *e;
-  /*! Flags to monitor. OTHER is only allowed if the socket is ARES_SOCKET_BAD.
-   */
-  ares_event_flags_t      flags;
-  /*! Callback to be called when event is triggered */
-  ares_event_cb_t         cb;
-  /*! Socket to monitor, allowed to be ARES_SOCKET_BAD if not monitoring a
-   *  socket. */
-  ares_socket_t           fd;
-  /*! Data associated with event handle that will be passed to the callback.
-   *  Optional, may be NULL. */
-  /*! Data to be passed to callback. Optional, may be NULL. */
-  void                   *data;
-  /*! When cleaning up the registered event (either when removed or during
-   *  shutdown), this function will be called to clean up the user-supplied
-   *  data. Optional, May be NULL. */
-  ares_event_free_data_t  free_data_cb;
-  /*! Callback to call to trigger an event. */
-  ares_event_signal_cb_t  signal_cb;
-};
-
-
-typedef struct {
-  const char   *name;
-  ares_bool_t (*init)(ares_event_thread_t *e);
-  void        (*destroy)(ares_event_thread_t *e);
-  void        (*event_add)(ares_event_thread_t *e, ares_event_t *event);
-  void        (*event_del)(ares_event_thread_t *e, ares_event_t *event);
-  void        (*event_mod)(ares_event_thread_t *e, ares_event_t *event,
-                           ares_event_flags_t new_flags);
-  size_t      (*wait)(ares_event_thread_t *e, unsigned long timeout_ms);
-} ares_event_sys_t;
-
-
-struct ares_event_thread {
-  /*! Whether the event thread should be online or not.  Checked on every wake
-   *  event before sleeping. */
-  ares_bool_t             isup;
-  /*! Handle to the thread for joining during shutdown */
-  ares__thread_t         *thread;
-  /*! Lock to protect the data contained within the event thread itself */
-  ares__thread_mutex_t   *mutex;
-  /*! Reference to the ares channel, for being able to call things like
-   *  ares_timeout() and ares_process_fd(). */
-  ares_channel_t         *channel;
-  /*! Not-yet-processed event handle updates.  These will get enqueued by a
-   *  thread other than the event thread itself. The event thread will then
-   *  be woken then process these updates itself */
-  ares__llist_t          *ev_updates;
-  /*! Registered event handles. */
-  ares__htable_asvp_t    *ev_handles;
-  /*! Pointer to the event handle which is used to signal and wake the event
-   *  thread itself.  This is needed to be able to do things like update the
-   *  file descriptors being waited on and to wake the event subsystem during
-   *  shutdown */
-  ares_event_t           *ev_signal;
-  /* Event subsystem callbacks */
-  const ares_event_sys_t *ev_sys;
-  /* Event subsystem private data */
-  void                   *ev_sys_data;
-};
+#include "ares_event.h"
 
 static void ares_event_destroy_cb(void *arg)
 {
@@ -147,41 +48,15 @@ static void ares_event_destroy_cb(void *arg)
   ares_free(event);
 }
 
-/*! Queue an update for the event handle.  Will search by the fd passed if
- *  not ARES_SOCKET_BAD to find a match and perform an update or delete
- *  (depending on flags).  Otherwise will add.  Do not use the event handle
- *  returned if its not guaranteed to be an add operation.
- *
- *  \param[out] event        Event handle. Optional, can be NULL.  This handle
- *                           will be invalidate quickly if the result of the
- *                           operation is not an ADD.
- *  \param[in]  e            pointer to event thread handle
- *  \param[in]  flags        flags for the event handle.  Use ARES_EVENT_FLAG_NONE
- *                           if removing a socket from queue (not valid if socket
- *                           is ARES_SOCKET_BAD).  Non-socket events cannot be
- *                           removed, and must have ARES_EVENT_FLAG_OTHER set.
- *  \param[in]  cb           Callback to call when event is triggered. Required.
- *                           Not allowed to be changed, ignored on modification.
- *  \param[in]  fd           File descriptor/socket to monitor.  May be
- *                           ARES_SOCKET_BAD if not monitoring file descriptor.
- *  \param[in]  data         Optional. Caller-supplied data to be passed to
- *                           callback. Only allowed on initial add, cannot be
- *                           modified later, ignored on modification.
- *  \param[in]  free_data_cb Optional. Callback to clean up caller-supplied
- *                           data. Only allowed on initial add, cannot be
- *                           modified later, ignored on modification.
- *  \param[in]  signal_cb    Optional. Callback to call to trigger an event.
- *
- *  \return ARES_SUCCESS on success
- */
-static ares_status_t ares_event_update(ares_event_t       **event,
-                                       ares_event_thread_t *e,
-                                       ares_event_flags_t  flags,
-                                       ares_event_cb_t     cb,
-                                       ares_socket_t       fd,
-                                       void               *data,
-                                       ares_event_free_data_t free_data_cb,
-                                       ares_event_signal_cb_t signal_cb)
+
+ares_status_t ares_event_update(ares_event_t       **event,
+                                ares_event_thread_t *e,
+                                ares_event_flags_t  flags,
+                                ares_event_cb_t     cb,
+                                ares_socket_t       fd,
+                                void               *data,
+                                ares_event_free_data_t free_data_cb,
+                                ares_event_signal_cb_t signal_cb)
 {
   ares_event_t *ev = NULL;
 
@@ -235,432 +110,6 @@ static void ares_event_signal(const ares_event_t *event)
     return;
   event->signal_cb(event);
 }
-
-
-#ifdef HAVE_PIPE
-typedef struct {
-  int filedes[2];
-} ares_pipeevent_t;
-
-static void ares_pipeevent_destroy(ares_pipeevent_t *p)
-{
-  if (p->filedes[0] != -1)
-    close(p->filedes[0]);
-  if (p->filedes[1] != -1)
-    close(p->filedes[1]);
-
-  ares_free(p);
-}
-
-static void ares_pipeevent_destroy_cb(void *arg)
-{
-  ares_pipeevent_destroy(arg);
-}
-
-static ares_pipeevent_t *ares_pipeevent_init(void)
-{
-  ares_pipeevent_t *p = ares_malloc_zero(sizeof(*p));
-  if (p == NULL)
-    return NULL;
-
-  p->filedes[0] = -1;
-  p->filedes[1] = -1;
-
-#  ifdef HAVE_PIPE2
-  if (pipe2(p->filedes, O_NONBLOCK|O_CLOEXEC) != 0) {
-    ares_pipeevent_destroy(p);
-    return NULL;
-  }
-#  else
-  if (pipe(p->filedes) != 0) {
-    ares_pipeevent_destroy(p);
-    return NULL;
-  }
-
-#    ifdef O_NONBLOCK
-  {
-    int val;
-    val = fcntl(p->filedes[0], F_GETFL, 0);
-    if (val >= 0) {
-      val |= O_NONBLOCK;
-    }
-    fcntl(p->filedes[0], F_SETFL, val);
-
-    val = fcntl(p->filedes[1], F_GETFL, 0);
-    if (val >= 0) {
-      val |= O_NONBLOCK;
-    }
-    fcntl(p->filedes[1], F_SETFL, val);
-  }
-#    endif
-
-#    ifdef O_CLOEXEC
-  fcntl(p->filedes[0], F_SETFD, O_CLOEXEC);
-  fcntl(p->filedes[1], F_SETFD, O_CLOEXEC);
-#    endif
-#endif
-
-#  ifdef F_SETNOSIGPIPE
-  fcntl(p->filedes[0], F_SETNOSIGPIPE, 1);
-  fcntl(p->filedes[1], F_SETNOSIGPIPE, 1);
-#  endif
-
-  return p;
-}
-
-
-static void ares_pipeevent_signal(const ares_event_t *e)
-{
-  ares_pipeevent_t *p;
-
-  if (e == NULL || e->data == NULL)
-    return;
-
-  p = e->data;
-  write(p->filedes[1], "1", 1);
-}
-
-static void ares_pipeevent_cb(ares_event_thread_t *e, ares_socket_t fd,
-                              void *data, ares_event_flags_t flags)
-{
-  unsigned char     buf[32];
-  ares_pipeevent_t *p = NULL;
-
-  (void)e;
-  (void)fd;
-  (void)flags;
-
-  if (data == NULL)
-    return;
-
-  p = data;
-
-  while (read(p->filedes[0], buf, sizeof(buf)) == sizeof(buf)) {
-    /* Do nothing */
-  }
-
-}
-
-
-static ares_event_t *ares_pipeevent_create(ares_event_thread_t *e)
-{
-  ares_event_t     *event = NULL;
-  ares_pipeevent_t *p     = NULL;
-  ares_status_t     status;
-
-  p = ares_pipeevent_init();
-  if (p == NULL)
-    return NULL;
-
-  status = ares_event_update(&event, e, ARES_EVENT_FLAG_READ,
-                             ares_pipeevent_cb,
-                             p->filedes[0],
-                             p,
-                             ares_pipeevent_destroy_cb,
-                             ares_pipeevent_signal);
-  if (status != ARES_SUCCESS) {
-    ares_pipeevent_destroy(p);
-    return NULL;
-  }
-
-  return event;
-}
-
-#endif
-
-
-#if defined(HAVE_POLL)
-
-static ares_bool_t ares_evsys_poll_init(ares_event_thread_t *e)
-{
-  e->ev_signal = ares_pipeevent_create(e);
-  if (e->ev_signal == NULL)
-    return ARES_FALSE;
-  return ARES_TRUE;
-}
-
-static void ares_evsys_poll_destroy(ares_event_thread_t *e)
-{
-  (void)e;
-}
-
-static void ares_evsys_poll_event_add(ares_event_thread_t *e, ares_event_t *event)
-{
-  (void)e;
-  (void)event;
-}
-
-static void ares_evsys_poll_event_del(ares_event_thread_t *e, ares_event_t *event)
-{
-  (void)e;
-  (void)event;
-}
-
-static void ares_evsys_poll_event_mod(ares_event_thread_t *e, ares_event_t *event, ares_event_flags_t new_flags)
-{
-  (void)e;
-  (void)event;
-  (void)new_flags;
-}
-
-static size_t ares_evsys_poll_wait(ares_event_thread_t *e, unsigned long timeout_ms)
-{
-  size_t            num_fds = 0;
-  ares_socket_t    *fdlist  = ares__htable_asvp_keys(e->ev_handles, &num_fds);
-  struct pollfd    *pollfd  = NULL;
-  int               rv;
-  size_t            cnt     = 0;
-  size_t            i;
-
-  if (num_fds) {
-    pollfd = ares_malloc_zero(sizeof(*pollfd) * num_fds);
-    for (i=0; i<num_fds; i++) {
-      ares_event_t *ev = ares__htable_asvp_get_direct(e->ev_handles, fdlist[i]);
-      pollfd[i].fd     = ev->fd;
-      if (ev->flags & ARES_EVENT_FLAG_READ)
-        pollfd[i].events |= POLLIN;
-      if (ev->flags & ARES_EVENT_FLAG_WRITE)
-        pollfd[i].events |= POLLOUT;
-    }
-  }
-  ares_free(fdlist);
-
-  rv = poll(pollfd, (nfds_t)num_fds, (timeout_ms == 0)?-1:(int)timeout_ms);
-  if (rv <= 0)
-    goto done;
-
-  for (i=0; i<num_fds; i++) {
-    ares_event_t      *ev;
-    ares_event_flags_t flags = 0;
-
-    if (pollfd[i].revents == 0)
-      continue;
-
-    cnt++;
-
-    ev = ares__htable_asvp_get_direct(e->ev_handles, pollfd[i].fd);
-    if (ev == NULL || ev->cb == NULL)
-      continue;
-
-    if (pollfd[i].revents & (POLLERR|POLLHUP|POLLIN))
-      flags |= ARES_EVENT_FLAG_READ;
-
-    if (pollfd[i].revents & POLLOUT)
-      flags |= ARES_EVENT_FLAG_WRITE;
-
-    ev->cb(e, pollfd[i].fd, ev->data, flags);
-  }
-
-done:
-  ares_free(pollfd);
-  return cnt;
-}
-
-
-static const ares_event_sys_t ares_evsys_poll = {
-  "poll",
-  ares_evsys_poll_init,
-  ares_evsys_poll_destroy,   /* NoOp */
-  ares_evsys_poll_event_add, /* NoOp */
-  ares_evsys_poll_event_del, /* NoOp */
-  ares_evsys_poll_event_mod, /* NoOp */
-  ares_evsys_poll_wait
-};
-
-#endif
-
-#ifdef HAVE_KQUEUE
-
-typedef struct {
-  int            kqueue_fd;
-  struct kevent *changelist;
-  size_t         nchanges;
-  size_t         nchanges_alloc;
-} ares_evsys_kqueue_t;
-
-static void ares_evsys_kqueue_destroy(ares_event_thread_t *e)
-{
-  ares_evsys_kqueue_t *kq = NULL;
-
-  if (e == NULL)
-    return;
-
-  kq = e->ev_sys_data;
-  if (kq == NULL)
-    return;
-
-  if (kq->kqueue_fd != -1) {
-    close(kq->kqueue_fd);
-  }
-
-  ares_free(kq->changelist);
-  ares_free(kq);
-  e->ev_sys_data = NULL;
-}
-
-static ares_bool_t ares_evsys_kqueue_init(ares_event_thread_t *e)
-{
-  ares_evsys_kqueue_t *kq = NULL;
-
-  kq = ares_malloc_zero(sizeof(*kq));
-  if (kq == NULL) {
-    return ARES_FALSE;
-  }
-
-  e->ev_sys_data = kq;
-
-  kq->kqueue_fd = kqueue();
-  if (kq->kqueue_fd == -1) {
-    ares_evsys_kqueue_destroy(e);
-    return ARES_FALSE;
-  }
-
-  kq->nchanges_alloc = 8;
-  kq->changelist     = ares_malloc_zero(sizeof(*kq->changelist) *
-                                        kq->nchanges_alloc);
-  if (kq->changelist == NULL) {
-    ares_evsys_kqueue_destroy(e);
-    return ARES_FALSE;
-  }
-
-  e->ev_signal = ares_pipeevent_create(e);
-  if (e->ev_signal == NULL) {
-    ares_evsys_kqueue_destroy(e);
-    return ARES_FALSE;
-  }
-
-  return ARES_TRUE;
-}
-
-static void ares_evsys_kqueue_enqueue(ares_evsys_kqueue_t *kq, int fd, int16_t filter, uint16_t flags)
-{
-  size_t idx;
-
-  if (kq == NULL)
-    return;
-
-  idx = kq->nchanges;
-
-  kq->nchanges++;
-
-  if (kq->nchanges > kq->nchanges_alloc) {
-    kq->nchanges_alloc <<= 1;
-    kq->changelist = ares_realloc_zero(kq->changelist, kq->nchanges_alloc >> 1,
-                                       kq->nchanges_alloc);
-  }
-
-  EV_SET(&kq->changelist[idx], fd, filter, flags, 0, 0, 0);
-}
-
-static void ares_evsys_kqueue_event_process(ares_event_thread_t *e, ares_event_t *event, ares_event_flags_t old_flags, ares_event_flags_t new_flags)
-{
-  ares_evsys_kqueue_t *kq;
-
-  if (e == NULL)
-    return;
-
-  kq = e->ev_sys_data;
-  if (kq == NULL)
-    return;
-
-  if (new_flags & ARES_EVENT_FLAG_READ &&
-      !(old_flags & ARES_EVENT_FLAG_READ)) {
-    ares_evsys_kqueue_enqueue(kq, event->fd, EVFILT_READ, EV_ADD | EV_ENABLE);
-  }
-
-  if (!(new_flags & ARES_EVENT_FLAG_READ) &&
-      old_flags & ARES_EVENT_FLAG_READ) {
-    ares_evsys_kqueue_enqueue(kq, event->fd, EVFILT_READ, EV_DELETE);
-  }
-
-  if (new_flags & ARES_EVENT_FLAG_WRITE &&
-      !(old_flags & ARES_EVENT_FLAG_WRITE)) {
-    ares_evsys_kqueue_enqueue(kq, event->fd, EVFILT_WRITE, EV_ADD | EV_ENABLE);
-  }
-
-  if (!(new_flags & ARES_EVENT_FLAG_WRITE) &&
-      old_flags & ARES_EVENT_FLAG_WRITE) {
-    ares_evsys_kqueue_enqueue(kq, event->fd, EVFILT_WRITE, EV_DELETE);
-  }
-}
-
-static void ares_evsys_kqueue_event_add(ares_event_thread_t *e, ares_event_t *event)
-{
-  ares_evsys_kqueue_event_process(e, event, 0, event->flags);
-}
-
-static void ares_evsys_kqueue_event_del(ares_event_thread_t *e, ares_event_t *event)
-{
-  ares_evsys_kqueue_event_process(e, event, event->flags, 0);
-}
-
-static void ares_evsys_kqueue_event_mod(ares_event_thread_t *e, ares_event_t *event, ares_event_flags_t new_flags)
-{
-  ares_evsys_kqueue_event_process(e, event, event->flags, new_flags);
-}
-
-
-static size_t ares_evsys_kqueue_wait(ares_event_thread_t *e, unsigned long timeout_ms)
-{
-  struct kevent        events[8];
-  size_t               nevents = sizeof(events) / sizeof(*events);
-  ares_evsys_kqueue_t *kq      = e->ev_sys_data;
-  int                  rv;
-  size_t               i;
-  struct timespec      ts;
-  struct timespec     *timeout = NULL;
-  size_t               cnt     = 0;
-
-  if (timeout_ms != 0) {
-    ts.tv_sec  = timeout_ms / 1000;
-    ts.tv_nsec = (timeout_ms % 1000) * 1000 * 1000;
-    timeout    = &ts;
-  }
-
-  memset(events, 0, sizeof(events));
-
-  rv = kevent(kq->kqueue_fd, kq->changelist, (int)kq->nchanges, events, (int)nevents, timeout);
-  if (rv < 0) {
-    return 0;
-  }
-
-  /* Changelist was consumed */
-  kq->nchanges = 0;
-  nevents      = (size_t)rv;
-
-  for (i=0; i<nevents; i++) {
-    ares_event_t      *ev;
-    ares_event_flags_t flags = 0;
-
-    ev = ares__htable_asvp_get_direct(e->ev_handles, (ares_socket_t)events[i].ident);
-    if (ev == NULL || ev->cb == NULL)
-      continue;
-
-    cnt++;
-
-    if (events[i].filter == EVFILT_READ || events[i].flags & (EV_EOF|EV_ERROR)) {
-      flags |= ARES_EVENT_FLAG_READ;
-    } else {
-      flags |= ARES_EVENT_FLAG_WRITE;
-    }
-
-    ev->cb(e, ev->fd, ev->data, flags);
-  }
-
-  return cnt;
-}
-
-
-static const ares_event_sys_t ares_evsys_kqueue = {
-  "kqueue",
-  ares_evsys_kqueue_init,
-  ares_evsys_kqueue_destroy,
-  ares_evsys_kqueue_event_add,
-  ares_evsys_kqueue_event_del,
-  ares_evsys_kqueue_event_mod,
-  ares_evsys_kqueue_wait
-};
-#endif
 
 
 static void ares_event_thread_wake(ares_event_thread_t *e)
